% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementRecog.R
\name{ElementRecog}
\alias{ElementRecog}
\title{ElementRecog is a function to identify COREs}
\usage{
ElementRecog(InputData, windowSize_Vec, peakNumMax, peakNumMin)
}
\arguments{
\item{InputData}{The input data as a table including chromosome regions
in which the first column is chromosome annotation, and second and third columns are start and ending positions.}

\item{windowSize_Vec}{Vector of window sizes orderes based on Order of CORE}

\item{peakNumMax}{Maximum order of COREs (e.g. maximum number of peaks within COREs)}

\item{peakNumMin}{Minimum order of COREs (e.g. minimum number of peaks within COREs)}
}
\description{
ElementRecog is a function to identify COREs
}
