% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowVec.R
\name{WindowVec}
\alias{WindowVec}
\title{WindowVec is a function to specify window size for each Order of COREs}
\usage{
WindowVec(InputData, peakNumMin)
}
\arguments{
\item{InputData}{The input data as a table including chromosome regions in
which the first column is chromosome annotation,  and second and third
columns are start and ending positions.}

\item{peakNumMin}{Minimum Order of COREs}
}
\value{
Vector of window sizes from Order 2 up to maximum Order of COREs
}
\description{
WindowVec is a function to specify window size for each Order of COREs
}
