% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CREAM.R
\name{CREAM}
\alias{CREAM}
\title{CREAM is the main function for CORE identification}
\usage{
CREAM(in_path, out_path, WScutoff = 1.5, MinLength = 1000, peakNumMin = 2)
}
\arguments{
\item{in_path}{Path to the input file (The file inclusing the functional
regions)
Note. You have to make sure that there is no overlapping regions within the
input file}

\item{out_path}{The path in which you want to store the identified COREs}

\item{WScutoff}{Threshold used to identify WS within distribution of maximum
distance between peaks for each order of CORE}

\item{MinLength}{Criteria for the minimum number of functional regions in the
input file}

\item{peakNumMin}{Minimum number of peaks for CORE identification}
}
\value{
Bed file including the identified COREs
}
\description{
CREAM is the main function for CORE identification
}
\examples{
CREAM(system.file("extdata", "A549_Chr21.bed", package = "CREAM"),
system.file("extdata", "A549_Chr21_COREs.bed", package = "CREAM"),
MinLength = 1000, peakNumMin = 2)
}
