% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMinFilt.R
\name{PeakMinFilt}
\alias{PeakMinFilt}
\title{PeakMinFilt is a function to filter the lowest Order of COREs which distance
between functional regions is close to the corresponding Window Size}
\usage{
PeakMinFilt(Clusters_init, WindowVecFinal)
}
\arguments{
\item{Clusters_init}{Table of indetified COREs before filteration}

\item{WindowVecFinal}{Vector of window sizes ordered based on order of CORE}
}
\value{
Minimum order of COREs
}
\description{
PeakMinFilt is a function to filter the lowest Order of COREs which distance
between functional regions is close to the corresponding Window Size
}
