% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowSizeRecog.R
\name{WindowSizeRecog}
\alias{WindowSizeRecog}
\title{WindowSizeRecog is a function to specify window size for each Order of COREs}
\usage{
WindowSizeRecog(InputData, COREorder)
}
\arguments{
\item{InputData}{The input data as a table including chromosome regions
in which the first column is chromosome annotation, and second and third
columns are start and ending positions.}

\item{COREorder}{Order of the COREs which window size has to be determined for.}
}
\value{
Window size identified for each order of CORE
}
\description{
WindowSizeRecog is a function to specify window size for each Order of COREs
}
